run more tests/improve printing/fix py reading/etc

+++ drone/src/drone/Input.cpp
@@ -1,14 +1,14 @@
 
 #include <drone/command/Input.hpp>
 
-double fourier_square(double t, double phase, double L, int n) {
-	double y = 0;
+float fourier_square(float t, float phase, float L, int n) {
+	float y = 0;
 	int j;
-	double t0 = phase / (2.0 * M_PI) * L;
+	float t0 = phase / (2.0 * M_PI) * L;
 	
 	for(int i = 0; i < n; ++i) {
 		j = 2 * i + 1;
-		y += sin((double)j * M_PI * (t+t0) / L) / (double)j;
+		y += sin((float)j * M_PI * (t+t0) / L) / (float)j;
 	}
 	y *= 4.0 / M_PI;
 	return y;
